Split off CrossTaskClassChangesIncrementalCompilationIntegrationTest
